Remove old scheduler
