--- apps/scheduler/src/exceptions/scheduler_error.py
@@ -1,2 +1,0 @@
-class SchedulerError(Exception):
-    pass
