--- apps/scheduler/src/models/location.py
@@ -1,8 +1,0 @@
-from dataclasses import dataclass
-from bson import ObjectId
-
-
-@dataclass
-class Location:
-    id: ObjectId
-    name: str
